requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aniso8601==9.0.1
async-timeout==4.0.2
attrs==21.4.0
cachetools==5.2.0
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
decorator==5.1.1
Flask==2.1.2
flask-restx==0.5.1
frozenlist==1.3.0
fsspec==2022.5.0
gcsfs==2022.5.0
google-api-core==2.8.2
google-auth==2.6.6
google-auth-oauthlib==0.5.2
google-cloud-core==2.3.2
google-cloud-run==0.2.0
google-cloud-storage==2.4.0
google-crc32c==1.3.0
google-resumable-media==2.3.3
googleapis-common-protos==1.56.4
grpc-google-iam-v1==0.12.4
grpcio==1.48.0
grpcio-status==1.48.0
idna==3.3
importlib-metadata==4.11.4
importlib-resources==5.7.1
iniconfig==1.1.1
isodate==0.6.1
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.5.1
kubernetes==23.6.0
lxml==4.9.0
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.5.5.64
packaging==21.3
pandas==1.4.2
Pillow==9.1.1
pluggy==1.0.0
proto-plus==1.20.6
protobuf==4.21.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycountry==22.3.5
pycryptodome==3.14.1
pyparsing==3.0.9
pyrsistent==0.18.1
PySocks==1.7.1
pytest==7.1.2
python-dateutil==2.8.2
python-dotenv==0.20.0
pytz==2022.1
PyYAML==6.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
six==1.16.0
slim==0.1
streamlink==4.1.0
tomli==2.0.1
urllib3==1.26.9
websocket-client==1.3.2
Werkzeug==2.1.2
yarl==1.7.2
zipp==3.8.0